+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Changelog
 * Runtime: fixed ``smoothed`` timestep policy not honoring average_cap during the very first frames.
 * Runtime: fixed a crash when managing empty units.
 * Runtime: Linux: fixed joypad D-pad.
+* Runtime: Linux: fixed ``caps_lock`` key detected as ``escape`` when released.
 * Data Compiler: fixed an issue that caused misdetected file changes before compilation.
 
 .. _v0.61.0: